+++ api/v1/views/users.py
@@ -8,7 +8,6 @@ from models import storage
 from models.user import User
 
 
-
 @app_views.route("/users/<uuid:user_id>", strict_slashes=False)
 def get_user(user_id):
     """Get user info"""
@@ -16,88 +15,32 @@ def get_user(user_id):
     user = storage.get(User, user_id)
     if not user:
         abort(404)
-    return jsonify(user.to_dict())
-
-# @app_views.route("/users", strict_slashes=False, methods=['POST'])
-# def new_user():
-#     """Create new user account"""
-#     if not request.is_json:
-#         abort(400, 'Not a JSON')
-#     data = request.get_json()
-#     if "first_name" not in data or "last_name" not in data\
-#             or "email" not in data or "password" not in data\
-#             or "birth_date" not in data:
-#                 abort(400, "Missing data")
-#     data['date_joined'] = str(date.today())
-#     if "gender" not in data:
-#         data['gender'] = 'U'
-#     if storage.check_user(data['email'], data['password']):
-#         return make_response(jsonify("The email or password has already been taken"), 409)
-#     new_user = User(**data)
-#     storage.add(new_user)
-#     storage.save()
-#     r = make_response(jsonify(new_user.to_dict()), 201)
-#     r.set_cookie("id", new_user.id)
-#     r.set_cookie("email", new_user.email)
-#     r.set_cookie("password", new_user.password)
-#     return r
+    user_repr = user.to_dict()
+    del user_repr['__class__']
+    del user_repr['password']
+    return jsonify(user_repr)
 
 @app_views.route("/users", strict_slashes=False, methods=['POST'])
 def new_user():
-    """Create a new user account"""
-    data = request.form
-    if "first_name" not in data or "last_name" not in data\
-            or "email" not in data or "password" not in data\
-            or "birth_date" not in data:
+     """Create new user account"""
+     if not request.is_json:
+         abort(400, 'Not a JSON')
+     data = request.get_json()
+     if "first_name" not in data or "last_name" not in data\
+             or "email" not in data or "password" not in data\
+             or "birth_date" not in data:
                 abort(400, "Missing data")
-    if "gender" not in data:
-        data['gender'] = 'U'
-    if storage.check_user(data['email'], data['password']):
-        return make_response(jsonify("The email or password has already been taken"), 409)
-    new_user = User(**data, date_joined=date.today())
-    storage.add(new_user)
-    storage.save()
-    r = make_response("Account Created Successfully", 201)
-    r.set_cookie("id", new_user.id)
-    r.set_cookie("email", new_user.email)
-    r.set_cookie("password", new_user.password)
-    return r
+     data['date_joined'] = str(date.today())
+     if "gender" not in data:
+         data['gender'] = 'U'
+     if storage.check_user(data['email'], data['password']):
+         return make_response(jsonify("The email or password has already been taken"), 409)
+     new_user = User(**data)
+     storage.add(new_user)
+     storage.save()
+     r = make_response(jsonify(new_user.to_dict()), 201)
+     return r
 
-@app_views.route("/login", strict_slashes=False, methods=['POST'])
-def login():
-    """Handeling Login form"""
-    data = request.form
-    if "email" not in data or "password" not in data:
-        abort(400, "Missing data")
-    user = storage.check_user(data['email'], data['password'])
-    if not user:
-        abort(404)
-    r = make_response("You are logged in, you are now able to CRUD", 200)
-    r.set_cookie("status", "In")
-    r.set_cookie("id", user.id)
-    r.set_cookie("email", user.email)
-    r.set_cookie("password", user.password)
-    return r
-
-@app_views.route("/delete_account", strict_slashes=False, methods=['POST', 'GET'])
-def delete_account():
-    """Handles delete user account"""
-    data = request.form
-    if "email" not in data or "password" not in data:
-        return make_response("Missing data")
-    user = storage.check_user(data['email'], data['password'])
-    if not user:
-        make_response("No such user account", 404)
-    email = request.cookies.get("email", None)
-    password = request.cookies.get("password", None)
-    if not email or not password:
-        make_response("You are not the owner", 401)
-    storage.delete(user)
-    r = make_response("Account Deleted", 200)
-    r.set_cookie("id", "", expires=0)
-    r.set_cookie("email", "", expires=0)
-    r.set_cookie("password", "", expires=0)
-    return r
 
 @app_views.route("/users/<uuid:user_id>", strict_slashes=False, methods=['DELETE'])
 def remove_user(user_id):
@@ -106,6 +49,18 @@ def remove_user(user_id):
     user = storage.get(User, user_id)
     if not user:
         abort(404)
+    if not request.is_json:
+        abort(400, "Not a JSON")
+    credentials = request.get_json()
+    email = credentials.get("email", None)
+    password = credentials.get("password", None)
+    if not email or not password:
+        abort(400, "Missing credentials")
+    credential_user = storage.credential_user(email, password)
+    if not credential_user:
+        abort(400, "Wrong credentials")
+    if credential_user.id != user.id:
+        abort(401)
     storage.delete(user)
     return make_response(jsonify({}), 200)
 
@@ -114,21 +69,36 @@ def update_user(user_id):
     """Update user info"""
     user_id = str(user_id)
     user = storage.get(User, user_id)
-    if not request.is_json:
-        abort(400, "Not a JSON")
     if not user:
         abort(404)
+    if not request.is_json:
+        abort(400, "Not a JSON")
     data = request.get_json()
-    allowed = ["password", "first_name", "education", "bio"]
-    filtered_data = {}
-    for k, v in data.items():
-        if k not in allowed:
-            pass
-        else:
-            filtered_data[k] = v
-    user.update(**filtered_data)
-    return make_response(jsonify(user.to_dict()), 200)
-        
+    em = data.get("em", None)
+    pawd = data.get("pawd", None)
+    password = data.get("password", None)
+    first_name = data.get("first_name", None)
+    education = data.get("education", None)
+    bio = data.get("bio", None)
+    if not em or not pawd:
+        abort(400, "Missing credentials")
+    credential_user = storage.credential_user(em, pawd)
+    if not credential_user:
+        abort(400, "Wrong credentials")
+    if user_id != credential_user.id:
+            abort(401)
+    if password or first_name or education or bio:
+        allowed = ["password", "first_name", "education", "bio"]
+        filtered_data = {}
+        for k, v in data.items():
+            if k not in allowed:
+                pass
+            else:
+                filtered_data[k] = v
+        user.update(**filtered_data)
+        return make_response(jsonify(user.to_dict()), 200)
+    abort(400, "Missing data")
+
 
 @app_views.route("/users/<uuid:user_id>/questions", strict_slashes=False)
 def get_user_questions(user_id):
